WebIDL/ecmascript-binding/legacy-platform-object/Set.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Legacy platform objects [[Set]] method</title>
<link rel="help" href="https://heycam.github.io/webidl/#legacy-platform-object-set">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    test(function() {
        var element = document.createElement("div");
        element.appendChild(document.createElement("div"));
        element.childNodes["5"] = "foo";
        assert_equals(element.attributes.foo, undefined);
    }, "must not set the property value when assigning to a numeric property on an object which implements an indexed property getter but not a setter when not in strict mode");

    test(function() {
        "use strict";
        var element = document.createElement("div");
        element.appendChild(document.createElement("div"));
        assert_throws(new TypeError(), function() { element.childNodes["5"] = "foo"; });
    }, "must throw when assigning to a numeric property on an object which implements a indexed property getter but not a setter in strict mode");
    
    test(function() {
        var element = document.createElement("div");
        element.attributes.foo = "foo";
        assert_equals(element.attributes.foo, "foo");
    }, "must allow assigning to a property on an object which implements a named property getter but not a setter when not in strict mode");

    test(function() {
        "use strict";
        var element = document.createElement("div");
        element.attributes.foo = "foo";
        assert_equals(element.attributes.foo, "foo");
    }, "must allow assigning to a property on an object which implements a named property getter but not a setter in strict mode");
</script>
